+++ mutraction-dom/src/compile-jsx.ts
@@ -171,14 +171,6 @@ export default function(_: Babel): PluginObj {
                                     );
                                     break;
 
-                                case "diagnostic":
-                                    if (value)
-                                        throw path.buildCodeFrameError("mu:diagnostic does not take a value.");
-                                    staticPropsForRuntime.push(
-                                        t.objectProperty(t.stringLiteral("mu:diagnostic"), t.booleanLiteral(true), true /* computed */)
-                                    );
-                                    break;
-                                    
                                 default:
                                     throw path.buildCodeFrameError(`Unsupported mutraction JSX attribute ${ name.name.name }`);
                             }